feat: Enhance dashboard and product management features with improved filtering and logging
- Updated DashboardController to include optional filters for period and shift in the summary endpoint, enhancing data retrieval flexibility.
- Enhanced logging in DashboardService and ProductosController to provide detailed information about request parameters and processing, improving debugging capabilities.
- Modified CategoriasController to return all active categories when no search term is provided, streamlining category retrieval.
- Improved front-end components to support new filtering options and enhance user experience with better error handling and visual feedback.

Functionality implemented:
- Strengthened the dashboard and product management functionalities, contributing to a more user-friendly and efficient application experience.

+++ tests/Frontend/RestaurantePro.Web.Admin.UnitTests/Pages/ProductosPageTests.cs
@@ -30,7 +30,7 @@ public class ProductosPageTests : TestContext
     {
         _productosApiMock.Setup(x => x.ObtenerProductosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>()))
                         .ReturnsAsync(new PaginatedList<ProductoDto> { Items = new List<ProductoDto>() });
-        _productosApiMock.Setup(x => x.ObtenerProductosPaginadosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<Guid?>(), It.IsAny<bool>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<DateTime?>(), It.IsAny<DateTime?>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<string>(), It.IsAny<string>()))
+        _productosApiMock.Setup(x => x.ObtenerProductosPaginadosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<Guid?>(), It.IsAny<bool>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<DateTime?>(), It.IsAny<DateTime?>(), It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<string>(), It.IsAny<string>()))
                         .ReturnsAsync(new PaginatedList<ProductoDto> { Items = new List<ProductoDto>() });
         _productosApiMock.Setup(x => x.ObtenerCategoriasAsync())
                         .ReturnsAsync(new List<CategoriaProductoDto>());
@@ -48,7 +48,7 @@ public class ProductosPageTests : TestContext
                         .ReturnsAsync(new PaginatedList<ProductoDto> { Items = productos, TotalCount = 2, PageNumber = 1, PageSize = 10 });
         _productosApiMock.Setup(x => x.ObtenerCategoriasAsync())
                         .ReturnsAsync(new List<CategoriaProductoDto>());
-        _productosApiMock.Setup(x => x.ObtenerProductosPaginadosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<Guid?>(), It.IsAny<bool>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<DateTime?>(), It.IsAny<DateTime?>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<string>(), It.IsAny<string>()))
+        _productosApiMock.Setup(x => x.ObtenerProductosPaginadosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<Guid?>(), It.IsAny<bool>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<DateTime?>(), It.IsAny<DateTime?>(), It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<string>(), It.IsAny<string>()))
                         .ReturnsAsync(new PaginatedList<ProductoDto> { Items = productos, TotalCount = 2, PageNumber = 1, PageSize = 10 });
     }
 
@@ -146,7 +146,7 @@ public class ProductosPageTests : TestContext
         var component = RenderComponent<Productos>();
 
         // Assert
-        _productosApiMock.Verify(x => x.ObtenerProductosPaginadosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<Guid?>(), It.IsAny<bool>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<DateTime?>(), It.IsAny<DateTime?>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<string>(), It.IsAny<string>()), Times.AtLeastOnce);
+        _productosApiMock.Verify(x => x.ObtenerProductosPaginadosAsync(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<Guid?>(), It.IsAny<bool>(), It.IsAny<decimal?>(), It.IsAny<decimal?>(), It.IsAny<DateTime?>(), It.IsAny<DateTime?>(), It.IsAny<int?>(), It.IsAny<int?>(), It.IsAny<string>(), It.IsAny<string>()), Times.AtLeastOnce);
     }
 
     [Fact]
